// src/app/model/event-sync.ts
export class EventSync {
  id: string;
  created: string;
  technicalName: string;
  target: string;
  name: string;
  action: string;

  provisioningResult: string;
  accountName: string;
  attributeValue: string;
  operation: string;

  attributeName: string;
  previousValue: string;
  errors: string;
}